SPT-1969: Renamed ConfirmDetailsSubmissionLambda
Renamed ConfirmDetailsSubmissionLambda to PostConfirmDetailsHandler,
which now follows our preferred naming strategy.
