+++ openAPI/public-api.yaml
@@ -103,9 +103,9 @@ paths:
       x-amazon-apigateway-integration:
         httpMethod: POST
         uri:
-          Fn::Sub: arn:aws:apigateway:${AWS::Region}:lambda:path/2015-03-31/functions/${ConfirmDetailsSubmissionLambda.Arn}:live/invocations
+          Fn::Sub: arn:aws:apigateway:${AWS::Region}:lambda:path/2015-03-31/functions/${PostConfirmDetailsHandler.Arn}:live/invocations
         credentials:
-          Fn::GetAtt: GetConfirmDetailsLambdaRestApiRole.Arn
+          Fn::GetAtt: PostConfirmDetailsHandlerRestApiRole.Arn
         type: aws_proxy
   /error/unrecoverable:
     get:
