+++ src/handlers/post-confirm-details-handler/tests/post-confirm-details-handler.test.ts
@@ -1,6 +1,6 @@
 import { APIGatewayEventRequestContextWithAuthorizer, APIGatewayProxyEvent } from "aws-lambda";
 import { expect, it } from "vitest";
-import { lambdaHandler } from "../confirm-details-submission-handler";
+import { lambdaHandler } from "../post-confirm-details-handler";
 
 it("should return a 302 status code on a successful request", async () => {
   const event = createMockAPIGatewayProxyEvent({}, "redirectUri=https%3A%2F%2Fapi.example.com&state=test-state-id");
